moved strings from model to constants. Renamed some icons
