+++ extensions/sql-migration/src/constants/notebookPathHelper.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-import * as loc from '../models/strings';
+import * as loc from './strings';
 
 export class NotebookPathHelper {
 	private static context: vscode.ExtensionContext;
