+++ extensions/sql-migration/src/constants/strings.ts
