+++ extensions/sql-migration/src/dialog/dialogBase.ts
@@ -1,0 +1,5 @@
+import * as azdata from 'azdata';
+
+export class DialogBase {
+
+}
